Add process probe-by-config endpoint and fix probe stream types
- Add v3_process_post_probe for POST /api/v3/process/probe (probe a process
  config directly; the cluster sibling v3_cluster_process_post_probe already
  existed). Requires Core v16.20.0+.
- Add optional coreid query parameter to v3_cluster_process_post_probe
- Fix ProcessProbeStream fps/bitrate_kbps/duration_sec from int to float to
  match the Core API; fractional values no longer raise a validation error
- Add unit tests (request building + fractional-number parsing); update
  README and CHANGELOG

+++ core_client/base/models/v3/process_probe_stream.py
@@ -24,13 +24,13 @@ class ProcessProbeStream(BaseModel):
     }
     """
 
-    bitrate_kbps: int
+    bitrate_kbps: float
     channels: int
     codec: str
     coder: str
-    duration_sec: int
+    duration_sec: float
     format: str
-    fps: int
+    fps: float
     height: int
     index: int
     language: str
